Move AutoBackupWallet into wallet.cpp
