+++ src/wallet/walletdb.h
@@ -149,6 +149,4 @@ private:
 
 void ThreadFlushWalletDB(const std::string& strFile);
 
-bool AutoBackupWallet (CWallet* wallet, std::string strWalletFile, std::string& strBackupWarning, std::string& strBackupError);
-
 #endif // BITCOIN_WALLET_WALLETDB_H
